requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
requests
openai-whisper
torch
transformers
accelerate
piper-tts
python-multipart